fix: changes according to pr comments
