+++ internal/loadbalance/load_balancing_test.go
@@ -1,4 +1,4 @@
-package loadbalance_test
+package loadbalance
 
 import (
 	"testing"
